+++ frontend/src/pages/station/inventory/intake/[id].vue
@@ -1,0 +1,17 @@
+/*
+ *     SPDX-License-Identifier: AGPL-3.0-only
+ *
+ *     Copyright (C) RainbowDashLabs and Contributor
+ */
+<script setup lang="ts">
+import IntakeView from '~/views/stationview/inventory/IntakeView.vue'
+
+definePageMeta({
+  layout: 'station',
+  name: 'inventory-intake',
+})
+</script>
+
+<template>
+  <IntakeView />
+</template>